artifacts/api-server/requirements.txt
discord.py>=2.3.0
asyncpg>=0.29.0
google-generativeai>=0.8.0
aiohttp>=3.9.0
